feat: WAF method "is in" uses a dropdown; clearer manage heading
- Extract the country picker's behavior into a generic OptionSelect
  (searchable single/multi select over a fixed option list); CountrySelect
  is now a thin wrapper over it.
- The HTTP method field's "is in" / "is not in" operators now pick from a
  multi-select dropdown of the known methods (GET/POST/…), mirroring the
  country picker, instead of a free-text tag input.
- Manage page heading: "Firewall" -> "Firewall rules" with a clearer
  subtitle, distinguishing it from the metrics ("Firewall activity") view.
