+++ src/lib/components/WafConditionRow.svelte
@@ -2,6 +2,7 @@
 	import Select from '$lib/components/Select.svelte'
 	import TagInput from '$lib/components/TagInput.svelte'
 	import CountrySelect from '$lib/components/CountrySelect.svelte'
+	import OptionSelect from '$lib/components/OptionSelect.svelte'
 	import {
 		fields,
 		getField,
@@ -30,6 +31,11 @@
 	const operators = $derived(operatorsForField(fieldMeta))
 	const multi = $derived(isMultiOperator(condition.operator))
 	const isTls = $derived(fieldType === 'tls')
+	// A closed-set string field (the HTTP method) picks from its suggestions via a
+	// dropdown rather than free text — single value as a combobox, list membership
+	// as a multi-select (mirrors the country picker).
+	const isMethod = $derived(fieldMeta?.value === 'method')
+	const suggestionOptions = $derived((fieldMeta?.suggestions ?? []).map((s) => ({ value: s, label: s })))
 	// Free-text combobox (datalist) for equals/not_equals on suggestion-backed fields.
 	const useCombobox = $derived(
 		!!fieldMeta?.suggestions &&
@@ -101,6 +107,12 @@
 					{:else}
 						<CountrySelect bind:value={condition.value} id="waf-value" />
 					{/if}
+				</div>
+			{:else if isMethod && multi}
+				<div class="field">
+					<label for="waf-values">Methods</label>
+					<OptionSelect multi options={suggestionOptions} bind:tags={valuesList}
+						id="waf-values" placeholder="Select methods" emptyText="No matching method" />
 				</div>
 			{:else if multi}
 				<div class="field">
